Add new gt4py tests folders with tests for versions
